requirements.txt
torch
torchvision
opencv-python-headless
Pillow
streamlit
numpy
tqdm
gpt4all
streamlit-option-menu
